requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Automat==22.10.0
babel==2.16.0
beautifulsoup4==4.12.2
bitsandbytes==0.43.3
black==24.8.0
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
boltons==21.0.0
bracex==2.5
Brotli==1.1.0
bs4==0.0.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.0.0
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1694444244250/work
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.8
datasets==2.20.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
easyocr==1.7.1
efficientnet-pytorch==0.7.1
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
face==22.0.0
fastapi==0.112.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock @ file:///croot/filelock_1700591183607/work
fire==0.6.0
Flask==3.0.0
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.0
gguf==0.9.1
glom==22.1.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
googleapis-common-protos==1.63.2
gpustat==1.1.1
gradio==4.41.0
gradio_client==1.3.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
hyperlink==21.0.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
inflate64==1.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lark==1.2.2
lazy_loader==0.3
librosa==0.10.2.post1
llama-recipes==0.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.6
loralib==0.1.2
lxml==5.0.1
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mizani==0.13.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
munch==4.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy @ file:///croot/numpy_and_numpy_base_1701295038894/work/dist/numpy-1.26.2-cp310-cp310-linux_x86_64.whl#sha256=2ab675fa590076aa37cc29d18231416c01ea433c0e93be0da3cfd734170cfc6f
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.39.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.21.3
orjson==3.10.7
outlines==0.0.46
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.6
peft==0.12.0
pexpect==4.9.0
Pillow @ file:///croot/pillow_1696580024257/work
pkuseg==0.0.25
platformdirs==4.2.2
plotnine==0.14.1
pooch==1.8.2
pretrainedmodels==0.7.4
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.0
protobuf==4.25.4
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.22.0
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybcj==1.0.2
pyclipper==1.3.0.post5
pycountry==24.6.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyppmd==1.1.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.2.0
pyzstd==0.16.1
queuelib==1.6.2
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.5.7
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.5.1
scipy==1.11.4
Scrapy==2.11.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
semantic-version==2.10.0
semgrep==1.84.1
Send2Trash==1.8.3
sentencepiece==0.2.0
service-identity==23.1.0
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.4
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.9.26
tiktoken==0.7.0
timm==0.9.7
tinycss==0.4
tinycss2==1.3.0
tldextract==5.1.1
tokenize-rt==6.0.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.1.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
triton==3.0.0
Twisted==22.10.0
typer==0.12.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.20.0
vllm==0.5.5
vllm-flash-attn==2.6.1
w3lib==2.1.2
watchfiles==0.24.0
wcmatch==8.5.2
wcwidth==0.2.12
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xformers==0.0.27.post2
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.0
zope.interface==6.1
